Move ast modules to common ast package.
